Added methods required in Transitions
Side note: Don't call anything regarding PlayerPrefs from another class.
Import Transitions and envoke the methods of that class. Unity really
doesn't like it when you have multiple classes changing the values of
PlayerPrefs and debugging that is a nightmare. This way it is a bit
simpler and a lot easy to fix if something goes wrong.