Got UI element generation to work
Still need to configure the prefabs to delete the component when clicked
in the action panel. I will also need to tweak the layout of the action
panel so that it doesn't look weird when you add a button. Side Note:
Make sure you play the game at 16:9 since I really don't want to work on
supporting multiple resolutions at this moment. Maybe we can work that
in later if we have time.